+++ src/pkg/report/sync/population_density/domain/domain.go
@@ -4,11 +4,11 @@ type PopulationDensityParams struct {
 	OKTMO            int
 	Year             int
 	PopulationAmount int
-	LandArea         float32
+	LandArea         float64
 }
 
 type PopulationDensityReport struct {
 	OKTMO   int
 	Year    int
-	Density float32
+	Density float64
 }
